move SparseEncodingQueryBuilder to upper level path
