+++ src/test/java/org/opensearch/neuralsearch/query/SparseEncodingQueryIT.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.neuralsearch.query.sparse;
+package org.opensearch.neuralsearch.query;
 
 import static org.opensearch.neuralsearch.TestUtils.objectToFloat;
 
@@ -19,6 +19,7 @@ import org.opensearch.index.query.MatchAllQueryBuilder;
 import org.opensearch.index.query.MatchQueryBuilder;
 import org.opensearch.neuralsearch.TestUtils;
 import org.opensearch.neuralsearch.common.BaseSparseEncodingIT;
+import org.opensearch.neuralsearch.query.SparseEncodingQueryBuilder;
 
 public class SparseEncodingQueryIT extends BaseSparseEncodingIT {
     private static final String TEST_BASIC_INDEX_NAME = "test-sparse-basic-index";
